layout/style/test/chrome/test_moz_document_condition_text.html

<!doctype html><html><head><meta charset="utf-8">
<title>Test for Bug 814907 (@-moz-document cases)</title>
<!-- Split from test_condition_text.html.
     This is a chrome test because @-moz-document can only be used
     from user style sheets.  See bug 1035091. -->

<script type="application/javascript" src="/content/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" type="text/css" href="/content/tests/SimpleTest/test.css"/>
</head><body>
  <a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=814907">Mozilla Bug 814907</a>
  <p id="display"></p>
  <div id="content" style="display: none"></div>
<pre id="test">
<script type="application/javascript">
SimpleTest.waitForExplicitFinish();

var windowUtils =
    window.QueryInterface(Components.interfaces.nsIInterfaceRequestor)
          .getInterface(Components.interfaces.nsIDOMWindowUtils);

var sheetURI =
    Components.classes["@mozilla.org/network/io-service;1"]
    .getService(Components.interfaces.nsIIOService)
    .newURI("moz_document_condition_text.css", null, document.baseURIObject);

var sheet =
    Components.classes["@mozilla.org/content/style-sheet-service;1"]
    .getService(Components.interfaces.nsIStyleSheetService)
    .preloadSheet(sheetURI, Components.interfaces.nsIStyleSheetService.USER_SHEET);

windowUtils.addSheet(sheet, windowUtils.USER_SHEET);

addLoadEvent(function () {
  var conditions = [
    "url(\"http://www.example.com/\")",
    "url(\"http://www.example.com/\")",
    "url(\"http://www.example.com/\")",
    "url-prefix(\"http://www.example.com/\")",
    "url-prefix(\"http://www.example.com/\")",
    "domain(\"example.com\")",
    "domain(\"example.com\")",
    "regexp(\"http://www.w3.org/TR/\\\\d{4}/[^/]*-CSS2-\\\\d{8}/\")",
    "regexp(\"http://www.w3.org/TR/\\\\d{4}/[^/]*-CSS2-\\\\d{8}/\")",
  ];

  is(sheet.cssRules.length, conditions.length);

  for (var i = 0; i < sheet.cssRules.length; i++) {
    var rule = sheet.cssRules[i];
    is(rule.conditionText, conditions[i], "rule " + i + " has expected conditionText");
    if (rule.type == CSSRule.MEDIA_RULE) {
      is(rule.conditionText, rule.media.mediaText, "rule " + i + " conditionText matches media.mediaText");
    }
  }

  SimpleTest.finish();
});
</script>
</pre>
</body>
</html>
